extract redis configuration
