+++ src/main/java/com/meep/mobilityresources/config/AppConfiguration.java
@@ -3,27 +3,10 @@ package com.meep.mobilityresources.config;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
 public class AppConfiguration {
-
-  //Creating Connection with Redis
-  @Bean
-  public RedisConnectionFactory redisConnectionFactory() {
-    return new LettuceConnectionFactory();
-  }
-
-  //Creating RedisTemplate
-  @Bean
-  public RedisTemplate<String, Object> redisTemplate(){
-    RedisTemplate<String, Object> template = new RedisTemplate<>();
-    template.setConnectionFactory(redisConnectionFactory());
-    return template;
-  }
 
   @Bean
   public RestTemplate restTemplate(RestTemplateBuilder restTemplateBuilder) {
