+++ src/main/java/com/meep/mobilityresources/config/RedisConfiguration.java
@@ -1,0 +1,25 @@
+package com.meep.mobilityresources.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+
+@Configuration
+public class RedisConfiguration {
+
+  //Creating Connection with Redis
+  @Bean
+  public RedisConnectionFactory redisConnectionFactory() {
+    return new LettuceConnectionFactory();
+  }
+
+  //Creating RedisTemplate
+  @Bean
+  public RedisTemplate<String, Object> redisTemplate(){
+    RedisTemplate<String, Object> template = new RedisTemplate<>();
+    template.setConnectionFactory(redisConnectionFactory());
+    return template;
+  }
+}
